Refactor preview app to go into cloudcare app
