+++ corehq/apps/cloudcare/static/cloudcare/js/preview_app/dragscroll.js
