+++ corehq/apps/cloudcare/templates/preview_app/base.html
@@ -143,7 +143,7 @@
         {% endcompress %}
 
         {% compress js %}
-        <script src="{% static 'preview_app/js/dragscroll.js' %}"></script>
+        <script src="{% static 'cloudcare/js/preview_app/dragscroll.js' %}"></script>
         {% endcompress %}
 
         <style id="case-tiles-style"></style>
@@ -178,7 +178,7 @@
         <script>
         $(function () {
             window.GMAPS_API_KEY = '{{ maps_api_key|safe }}'; // maps api is loaded on-demand
-            hqImport('preview_app/js/preview_app.js').start({
+            hqImport('cloudcare/js/preview_app/preview_app.js').start({
                 apps: [{{ app|JSON }}],
                 language: "{{ request.user.language|default:'en' }}",
                 username: "{{ request.user.username  }}",
@@ -209,6 +209,6 @@
         {% include 'formplayer/progress.html' %}
         {% include 'formplayer/dependencies.html' %}
         {% include "style/includes/ko.html" %}
-        <script src="{% static 'preview_app/js/preview_app.js' %}"></script>
+        <script src="{% static 'cloudcare/js/preview_app/preview_app.js' %}"></script>
     </body>
 </html>
